requirements.txt
python-telegram-bot
pandas
python-dotenv
google-generativeai